Restructure plugin to remove complexity
